Install nfpm in Travis as part of the toolchain
We pin to nfpm version 0.9.4 since it has binary releases available for
both Darwin and Linux, which 0.9.5 curiously does not. Unfortunately
this version does not support the "${VERSION}" syntax, so we'll need to
transform the configuration file prior to calling, or find a better way
to distribute binaries for late versions of nfpm.