Don't error in `pulumi stack output` when there is no root resource
We used to issue an error when you ran `pulumi stack output` for a
stack with with no root resource (e.g. one that hadn't been stood up
yet or one that had been stood up but then destoryed).

Instead, just follow the normal case for a stack that has a root
resource, but has no outputs.

Fixes #2016