Add an API for importing stack outputs
These changes add a new API to the Pulumi SDK, `service.getStack`, that
returns the outputs of a given stack. The Pulumi account performing the
deployment that calls this API must have access to the indicated stack
or the call will fail.

This API is implemented as an invoke that is implemented by a builtin
provider managed by the engine. This provider will be used for any
custom resources and invokes inside the `pulumi:pulumi` module.
Currently this provider's API is exactly the `pulumi:pulumi:getStack`
invoke.

This is the short-term fix for #109.