Stop `yarn link`'ing the package under test in lifecycle tests
Instead of `yarn link`ing, pass an actual path to the library we just
built, so it is installed as a normal node module. This gives a better
indication of how things will work once the package is actually on
NPM.

It also makes it easier to in the future use more advanced yarn
features to do "testing at HEAD", because `yarn install` gets to see
the complete package graph.