Don't use a per test cache
The use of a per test cache is problematic right now because when run
on Node 11 it forces every test to build its own copy of gRPC from
source, which is very slow.

Go back to the shared cache (with --mutex network) but add a command
after running `yarn install` which cleans any local packages that were
installed. This is, unfortunately, the best we can do for now.