Store configuration files in `.pulumi` by default
Previously, we would write a stack's configuration file next to
Pulumi.yaml in a file named `Pulumi.<stack-name>.yaml`. This behavior
could be controled by setting the `config` property in the project's
`Pulumi.yaml` file.

This change updates the default storage location to be under a new
`.pulumi` folder that lives next to `Pulumi.yaml`. For configuration
files that are already present in the root, we'll continue to use
them, but configuration for new stacks will be created in `.pulumi`.

To go back to the old behavior, one can add `config: .` to their
Pulumi.yaml file.

Fixes #2005