Store stack specific config in folder named after owner
When the backend for a stack is the Pulumi Service, store
configuration values for the stack in a folder with the name of the
owner.

This means that `ellismg/my-great-stack` and `pulumi/my-great-stack`
now have different configuration files, instead of sharing the
`Pulumi.my-great-stack.yaml` file stored in the configuration root
directory. Since these are two different deployments, it makes sense
that they can be configured independently. In addition, since
encrypted configuration values are stored using a per stack key,
sharing the files doesn't actually work when using encrypted
configuration.

For compat, the CLI prefers the older file name when present.

Fixes #1859