Lift dependencies from Output<Output>s
Today, if the callback supplied to Output.apply returns an Output, the
inner Output's dependencies are dropped. This decision was intentional,
as it was originally believed that this scenario would be undesirable,
as it typically arises when resources are created inside of an apply
(which can prevent them from being displayed during a preview).

Unfortunately, this causes dependencies to be dropped in scenarios such
as the Helm integration in `pulumi-kubernetes`, where the set of
resources to be created must be dependent on Input<>s.

This changes update the implementation of Output.apply to lift the
dependencies from any inner output into the result's resources. The code
takes the same approach used by isKnown.