More helpful toString on Output
It is common for users to intentionally or accidentally try to `toSting` an Output expecting to get the underlying value.  Instead, they currently get `[Object object]` with no clear reason why.

We can't give the actual underlying value because it is only asynchronously availabe, but we can provide a more helpful placeholder value with a pointer to documentation about how to work with Outputs.

Part of #2206.