Only show upgrade message when installed from get.pulumi.com
In general, we can't know what the latest version of `pulumi` in a
package manager might be and we've seen in practice that `brew` can
often take some time to pick up a new version.

For now, we'll just notify users there's a newer version present when
we are pretty confident it was installed from get.pulumi.com. For
folks installing `pulumi` from their package manager, we'll just let
their package manager advertise an update.